+++ src/components/input.js
@@ -30,9 +30,8 @@
     /^(-|\+)?((\d*(\.\d*)?)|((((\d+(\.\d*)?)|(\.\d+))(e(-|\+)?\d*)?)?))$/i;
   const integerRegExp = /^(-|\+)?\d+$/;
   const partialIntegerRegExp = /^(-|\+)?\d*$/;
-
-  // other than text type
-  const isAllowedHtmlInputType = (type) => type === 'radio' || type === 'checkbox';
+  const allowedHtmlInputTypes = ['text', 'password', 'email', 'url', 'tel', 'search', 'hidden',
+    'radio', 'checkbox'];
 
   const isBoolHtmlInputType = (type) => type === 'radio' || type === 'checkbox';
 
@@ -52,7 +51,9 @@
 
     string: (value) => value.trim(),
 
-    text: (value) => valuesNormalizer.string(value)
+    text: (value) => valuesNormalizer.string(value),
+
+    textarea: (value) => valuesNormalizer.string(value)
   };
 
   // type guards disallow unexpected characters in the input
@@ -89,7 +90,9 @@
       return true;
     },
 
-    text: (value, oldValue) => valueChangeGuards.string(value, oldValue)
+    text: (value, oldValue) => valueChangeGuards.string(value, oldValue),
+
+    textarea: (value, oldValue) => valueChangeGuards.string(value, oldValue)
   };
 
   class Input extends Component {
@@ -98,6 +101,7 @@
 
       this.state = {
         value: this.props.value,
+        // last normalized value
         oldValue: this.props.value
       };
 
@@ -113,7 +117,7 @@
 
       // we must ignore the same value when we have an intermediary string representation
       // like 3.20 (after the 0 there can be a new decimal)
-      if (!newProps.isInputPending && newProps.value !== this.props.value) {
+      if (!newProps.isInputPending && newProps.value !== this.state.oldValue) {
         this.setState({
           value: newProps.value,
           oldValue: newProps.value
@@ -178,19 +182,23 @@
 
     render() {
       const {errors, type, required, placeholder, className, onBlur, readOnly,
-        disabled, autoFocus} = this.props;
-      let {value} = this.state;
-
-      value = isVal(value)
-        ? value
+        disabled, autoFocus, inputClassName, inputProps} = this.props;
+      const value = isVal(this.state.value)
+        ? this.state.value
         : '';
+      const hasErrors = !!errors && !!errors.length;
 
-      return React.DOM.span({className},
+      return React.DOM.span({
+          className: `crizmas-input ${hasErrors ? 'has-errors' : ''} ${className}`
+        },
         !!required && React.DOM.span(null, '*'),
-        React.DOM.input({
+        React.DOM[type === 'textarea' ? type : 'input'](Object.assign(
+          {},
+          inputProps,
+          {
             value,
             checked: isBoolHtmlInputType(type) && value,
-            type: isAllowedHtmlInputType(type)
+            type: allowedHtmlInputTypes.includes(type)
               ? type
               : 'text',
             placeholder,
@@ -198,10 +206,11 @@
             onBlur,
             readOnly,
             disabled,
-            autoFocus
+            autoFocus,
+            className: inputClassName,
           }
-        ),
-        !!errors && !!errors.length && React.DOM.span(null,
+        )),
+        hasErrors && React.DOM.span(null,
           errors.map((error, i) => React.DOM.span({key: i}, error))
         )
       );
@@ -211,23 +220,26 @@
   Input.propTypes = {
     value: PropTypes.any,
     initialValue: PropTypes.any,
-    type: PropTypes.oneOf(['string', 'text', 'number', 'integer', 'checkbox', 'radio']),
+    type: PropTypes.oneOf([...allowedHtmlInputTypes, 'string', 'number','integer', 'textarea']),
     errors: PropTypes.array,
     isInputPending: PropTypes.bool,
     required: PropTypes.bool,
     placeholder: PropTypes.string,
     className: PropTypes.string,
+    inputClassName: PropTypes.string,
     debounce: PropTypes.number,
     onChange: PropTypes.func,
     onBlur: PropTypes.func,
     onStartPending: PropTypes.func,
     readOnly: PropTypes.bool,
     disabled: PropTypes.bool,
-    autoFocus: PropTypes.bool
+    autoFocus: PropTypes.bool,
+    inputProps: PropTypes.object
   };
 
   Input.defaultProps = {
-    type: 'string'
+    type: 'text',
+    className: ''
   };
 
   Input.contextTypes = {